refactor(stock) drag and drop in a separate file
